Abstracted some patterns
