+++ src/format/stats.ts
@@ -1,20 +1,13 @@
-import { matchesTranslation } from '../translate/match';
-import printf from '../translate/printf';
+import translate, { NO_DESCRIPTION, Stat } from '../translate';
 import {
   Description,
   Descriptions,
   StatLocaleData,
-  StatLocaleDatas,
-  Translation
+  StatLocaleDatas
 } from '../types/StatDescription';
 
 // arg types
-export type Stat = {
-  id: string;
-  value: number | [number, number];
-  alias?: string;
-};
-
+export { Stat } from '../translate';
 export type Options = {
   datas: StatLocaleDatas;
   fallback: Fallback | FallbackCallback;
@@ -88,8 +81,6 @@ function findDescription(stat_id: string, locale_data: Descriptions) {
   );
 }
 
-const NO_DESCRIPTION = 'NO_DESCRIPTION';
-
 // stats will get mutated
 function formatWithFinder(
   stats: Map<string, Stat>,
@@ -127,54 +118,6 @@ function formatWithFinder(
   }
 
   return lines;
-}
-
-function translate(
-  description: Description,
-  provided: Map<string, Stat>
-): string | undefined {
-  const { stats, no_description, translations } = description;
-
-  if (no_description === true) {
-    return NO_DESCRIPTION;
-  }
-
-  // intersect the required stat_ids from the desc with the provided
-  const required_stats = stats
-    .map(stat_id => {
-      const stat = provided.get(stat_id);
-
-      // default the value to 0
-      if (stat === undefined) {
-        return {
-          id: stat_id,
-          value: 0
-        };
-      } else {
-        return stat;
-      }
-    })
-    .filter((stat: Stat | null): stat is Stat => stat !== null);
-
-  const translation = matchingTranslation(translations, required_stats);
-
-  if (translation === undefined) {
-    return undefined;
-  } else {
-    return printf(
-      translation.text,
-      required_stats.map(({ value }) => value),
-      translation.formatters
-    );
-  }
-}
-
-function matchingTranslation(translations: Translation[], stats: Stat[]) {
-  const args = stats.map(({ value }) => value);
-
-  return translations.find(translation => {
-    return matchesTranslation(translation, args);
-  });
 }
 
 function formatWithFallback(
