+++ src/translate/index.ts
@@ -1,0 +1,6 @@
+import translate from './translate';
+export default translate;
+
+export { default as asRegexp } from './asRegexp';
+export { matchesTranslation } from './match';
+export { NO_DESCRIPTION, Stat } from './translate';
